Move TractionQcMessage to the Traction module
