+++ tol_lab_share/message_properties/definitions/labware.py
@@ -9,7 +9,7 @@ from tol_lab_share.message_properties.definitions.labware_type import LabwareTyp
 from tol_lab_share.message_properties.definitions.sample import Sample
 from tol_lab_share.messages.output_feedback_message import OutputFeedbackMessage
 from tol_lab_share.traction.output_traction_message import OutputTractionMessage
-from tol_lab_share.messages.traction_qc_message import TractionQcMessage
+from tol_lab_share.traction.traction_qc_message import TractionQcMessage
 
 from .message_property import MessageProperty
 from functools import singledispatchmethod
